HAI-160: refactored packages and moved Application (was HelloApplication) to correct place
