+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/Application.kt
@@ -1,11 +1,9 @@
-package fi.hel.haitaton.hello
+package fi.hel.haitaton.hanke
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
-import org.springframework.context.annotation.ComponentScan
 
 @SpringBootApplication
-@ComponentScan("fi.hel.haitaton")
 class HelloApplication
 
 fun main(args: Array<String>) {
