+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/hello/HelloHtmlController.kt
@@ -1,4 +1,4 @@
-package fi.hel.haitaton.hello
+package fi.hel.haitaton.hanke.hello
 
 import org.springframework.http.MediaType
 import org.springframework.stereotype.Controller
