+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/hello/HelloResponse.kt
@@ -1,4 +1,4 @@
-package fi.hel.haitaton.hello
+package fi.hel.haitaton.hanke.hello
 
 /**
  * A simple REST response payload for the HelloApiController.
